requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
pydantic==2.6.1
python-dotenv==1.0.1
openai>=1.12.0,<2.0.0
python-multipart==0.0.9
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
numpy
faiss-cpu
PyPDF2==3.0.1
docx2txt==0.8
python-docx==1.1.0
python-magic-bin==0.4.14
tiktoken==0.6.0 
mysqlclient==2.2.4
pymysql==1.1.0
psycopg2-binary==2.9.9